Refactor code and implement lombok
