+++ questionbank/src/main/java/com/example/questionbank/repository/LoadDatabase.java
@@ -1,11 +1,11 @@
-package questionbank.database;
+package com.example.questionbank.repository;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import questionbank.model.Question;
+import com.example.questionbank.model.Question;
 
 import java.util.Arrays;
