+++ questionbank/src/main/java/com/example/questionbank/repository/QuestionRepository.java
@@ -1,7 +1,7 @@
-package questionbank.database;
+package com.example.questionbank.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import questionbank.model.Question;
+import com.example.questionbank.model.Question;
 
 public interface QuestionRepository extends JpaRepository<Question, Long> {
